+++ HISTORY.rst
@@ -12,6 +12,9 @@ Release History
   querying ``Response.is_redirect`` and ``Response.headers['location']``.
   Advanced users will be able to process malformed redirects more easily.
 
+- Altered how ``SessionRedirectMixin.resolve_redirects`` and ``Session.send``
+  process redirect history.
+
 2.13.0 (2017-01-24)
 +++++++++++++++++++